A few more ecmodule updates

+++ schemes/pksig_ecdsa.py
@@ -35,7 +35,7 @@ class ECDSA(PKSig):
             r = group.zr(pk['g'] ** k)
             e = group.hash(M)
             s = (~k) * (e + x * r)
-            if (r == long(0) or s == long(0)):
+            if (r == 0 or s == 0):
                 print ("unlikely error r = %s, s = %s" % (r,s))
                 continue
             else:
